+++ build.config.ts
@@ -1,7 +1,3 @@
 import { defineBuildConfig } from 'unbuild';
 
-export default defineBuildConfig({
-  externals: [
-    'eslint-flat-config-utils',
-  ],
-});
+export default defineBuildConfig({});