+++ src/configs/index.ts
@@ -1,7 +1,7 @@
 export type {
   Config,
   Rules,
-} from '../core/types';
+} from '../core';
 
 export * from './css';
 export * from './eslint';